requirements.txt
fastapi
uvicorn
openai
sentence-transformers
faiss-cpu
requests
pytest
